Sort out use of '%' in categories.
This allows the example below through.

Note that 'F D' needs to matched to 'Algebra D'.  We know that 'F D' is
an integral domain, so is an algebra over itself, thus it should be
fine and dandy.

Algebra(T: with): Category == with;
IntegralDomain: Category == Algebra(%) with;
LocalAlgebra(R: with, S: Algebra R): with == add;

F(A: with): IntegralDomain with == add;
D: with == add;