Handle lists of records and similar in exports of types.
In particular, use the Syme on an incoming bit of absyn in tfPending to help decide if it is
'special' or not.

In tfPending, tform.c tries to convert absyn objects into TForms.  Most of the
time, it can be done by syntax (a AB_With corresponds to a TF_With, etc),
but for 'special' types like Record, Union and suchlike it needs to be done
by looking at available symbol meanings.

There's two cases of interest; when compiling, there is usually no symbol meaning
(syme) on the incoming absyn. Secondly, when inlining or substituting, we must
have a symbol meaning.  In this second case we want to use the syme
sitting on the incoming absyn.

The substitution case looks like:
List(T: with): with foo: % -> Generator T;
l: List Record(...) := ...;
foo(l)

The return type of foo(l) is Generator Record(...), which in absyn terms
looks like (Apply Foo (Apply Record ...)), and in tform terms
looks like (Generator (Record ...)).  This conversion requires that we recognise
that 'Record' refers to the magic built-in one.

If we happen to be working from a library, then we don't have an in-scope definition
of Record, but do have one on the incoming syme.

Lets give that syme a big warm hug and welcome it to the special case.