Fix warning where library symbol is not found.
The compiler picked the wrong branch of an editlevel #define.

Remove pointless, stupid and frankly daft use of editlevels.

Whatever wally thought they were a good idea needs to have things explained
to them over a nice cup of tea.  Perhaps with biscuits.

Maybe not biscuits.. Depends if they like them with tea or not.