Fix substitution of globals into inlined code
This mostly worked previously, except where we needed to know the type
of a global (generally on imports) introduced in the current inlining
phase.

Turns out that foamExprType wasn't smart enough to use the inlUnit's
list of globals, but referred back to the original prog.

Change to use the new foamExprTypeG0 function, which allows you to specify
a bunch of globals.