Save conditions off along with symes
Reasons for doing this are that on a substitution (such as an extend)
we replace the original tf with a 'with' and a list of symes.

We need to preserve conditional information on these symes as there's no
other way to recover it.  (aside from contextual, but that does not
work well, except for %%)