Add symeConditionalContext definition
When a syme is imported it takes its context from the importing type form -
for example, hash$Evalable S.  This context is the condition under which
the importing type is defined (in this case 'S has SetCategory').

Given that S satisfies SetCategory when inferring the type of
'Evalable S', we can compile it without hitting any type inference issues.