Cascaded imports fix
When importing from a type with cascaded exports, the tforms representing the
cascades must come from the current syntax tree.

This patch ensures that this is the case by replacing the imported type with
one returned from typeInferTForm()

Hopefully this should make things like 'export from T' more useful in contexts
like List T

Move the code into tinfer.c and ensure that tfQueries are inferred before the call
to stabImportFrom.

There's probably a subsequent stage to this where the cascades and queries
are inferred as a separate step, so we can deal with any interdependencies.