Fix "pretend" and comma interaction
(a, b) pretend X needs to make sure that the comma expression is converted to
to the representation form of X.. so, (a,b) pretend % should convert the multi
to a Cross.

Note to self about embed:  The type of an expression is always the "internal"
type.  If there is an embedding, then this information is held on the semantics
tab, and users are expected to take a look when assigning