sefo.c: track '%' on with expressions across file boundaries.
When we compile an expression like 'A has with f: %', a reference to % on
the 'with' expression wasn't being copied over across files.  So, now we
save it off in the semantics part of the abstract syntax, and ensure it
reappears when the tform for the expression is being computed.