types: Add support for java types.
The idea is that an import will be something along the idea of:

import Runnable: with { new: () -> %;
			(%, 'run') -> ()-> ();
} from Foreign Java "java.util";

This will introduce a type (Runnable), which can then be used as follows:

	r: Runnable := new();
	r.run()

Runnable will initially not work with normal types, so no List(Runnable),
but maybe later.

This assumes that a.b binds tighter than b(x), so that
a.b(x) is interpreted as (a.b)(x), so no need for new syntax.