Detect and use gmp.h
To work correctly on architectures with 64bit pointers and 32bit ints, we
need to import gmp.h as we can't currently have an 'int' type that works
in this case.

The major change here is that if gmp.h is not found at compile, then the
gmp related types will not be compiled.